Correction d'un bug avec la roulette après la mise à jour d'avant

+++ Java3D/src/fr/view/Panneau.java
@@ -112,7 +112,7 @@ public class Panneau extends JPanel {
 
 			this.addMouseWheelListener(new MouseWheelListener() {
 				public void mouseWheelMoved(MouseWheelEvent e) {
-					m.zoom((-e.getWheelRotation() + 15) / 15);
+					m.zoom((double)((-e.getWheelRotation() + 15.0)) / 15.0);
 					repaint();
 				}
 			});
